requirements.txt
pyyaml
Pillow
torchinfo
thop
tqdm
tensorboard
six
protobuf==3.20.3